reorganize react components
