+++ app/javascript/components/publicHealth/Import.js
