+++ app/javascript/components/publicHealth/PatientsTable.js
